requirements.txt
langchain 
transformers
torch
sqlalchemy 
requests 
pillow
faiss-cpu
datasets